+++ lib/flutter_adaptive_cards.dart
@@ -1,8 +1,3 @@
 library flutter_adaptive_cards;
 
 export 'package:flutter_adaptive_cards/src/flutter_adaptive_cards.dart';
-export 'package:flutter_adaptive_cards/src/basic_markdown.dart';
-export 'package:flutter_adaptive_cards/src/registry.dart';
-export 'package:flutter_adaptive_cards/src/elements/input.dart';
-export 'package:flutter_adaptive_cards/src/elements/actions.dart';
-export 'package:flutter_adaptive_cards/src/elements/basics.dart';
