+++ lib/src/elements/image.dart
@@ -1,0 +1,116 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_adaptive_cards/src/additional.dart';
+import 'package:flutter_adaptive_cards/src/base.dart';
+import 'package:flutter_adaptive_cards/src/utils.dart';
+
+class AdaptiveImage extends StatefulWidget with AdaptiveElementWidgetMixin {
+  AdaptiveImage({Key key, this.adaptiveMap}) : super(key: key);
+
+  final Map adaptiveMap;
+
+  @override
+  _AdaptiveImageState createState() => _AdaptiveImageState();
+}
+
+class _AdaptiveImageState extends State<AdaptiveImage> with AdaptiveElementMixin {
+  Alignment horizontalAlignment;
+  bool isPerson;
+  Tuple<double, double> size;
+
+  @override
+  void initState() {
+    super.initState();
+    horizontalAlignment = loadAlignment();
+    isPerson = loadIsPerson();
+    size = loadSize();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    //TODO alt text
+    Widget image = Image(
+      image: NetworkImage(url),
+      fit: BoxFit.contain,
+    );
+
+    if (isPerson) {
+      image = ClipOval(
+        clipper: FullCircleClipper(),
+        child: image,
+      );
+    }
+
+    image = Align(
+      alignment: horizontalAlignment,
+      child: image,
+    );
+
+    if (size != null) {
+      image = ConstrainedBox(
+        constraints: BoxConstraints(
+          minWidth: size.a,
+          minHeight: size.a,
+          maxHeight: size.b,
+          maxWidth: size.b,
+        ),
+        child: image,
+      );
+    }
+    return SeparatorElement(
+      adaptiveMap: adaptiveMap,
+      child: image,
+    );
+  }
+
+  Alignment loadAlignment() {
+    String alignmentString = adaptiveMap["horizontalAlignment"] ?? "left";
+    switch (alignmentString) {
+      case "left":
+        return Alignment.centerLeft;
+      case "center":
+        return Alignment.center;
+      case "right":
+        return Alignment.centerRight;
+      default:
+        return Alignment.centerLeft;
+    }
+  }
+
+  bool loadIsPerson() {
+    if (adaptiveMap["style"] == null || adaptiveMap["style"] == "default") return false;
+    return true;
+  }
+
+  String get url => adaptiveMap["url"];
+
+  Tuple<double, double> loadSize() {
+    String sizeDescription = adaptiveMap["size"] ?? "auto";
+    sizeDescription = sizeDescription.toLowerCase();
+
+    int size;
+    if (sizeDescription != "auto" && sizeDescription != "stretch") {
+      size = resolver.resolve("imageSizes", sizeDescription);
+    }
+
+    var width = size;
+    var height = size;
+
+    // Overwrite dynamic size if fixed size is given
+    if (adaptiveMap["width"] != null) {
+      var widthString = adaptiveMap["width"].toString();
+      widthString = widthString.substring(0, widthString.length - 2); // remove px
+      width = int.parse(widthString);
+    }
+    if (adaptiveMap["height"] != null) {
+      var heightString = adaptiveMap["height"].toString();
+      heightString = heightString.substring(0, heightString.length - 2); // remove px
+      height = int.parse(heightString);
+    }
+
+    if (height == null || width == null) {
+      return null;
+    }
+
+    return Tuple(width.toDouble(), height.toDouble());
+  }
+}
