+++ lib/src/elements/media.dart
@@ -1,0 +1,61 @@
+import 'package:chewie/chewie.dart';
+import 'package:flutter/material.dart';
+import 'package:video_player/video_player.dart';
+
+import '../additional.dart';
+import '../base.dart';
+import '../utils.dart';
+
+class AdaptiveMedia extends StatefulWidget with AdaptiveElementWidgetMixin {
+  AdaptiveMedia({Key key, this.adaptiveMap}) : super(key: key);
+
+  final Map adaptiveMap;
+
+  @override
+  _AdaptiveMediaState createState() => _AdaptiveMediaState();
+}
+
+class _AdaptiveMediaState extends State<AdaptiveMedia> with AdaptiveElementMixin {
+  VideoPlayerController videoPlayerController;
+  ChewieController controller;
+
+  String sourceUrl;
+  String postUrl;
+  String altText;
+
+  FadeAnimation imageFadeAnim = FadeAnimation(child: const Icon(Icons.play_arrow, size: 100.0));
+
+  @override
+  void initState() {
+    super.initState();
+
+    postUrl = adaptiveMap["poster"];
+    sourceUrl = adaptiveMap["sources"][0]["url"];
+    videoPlayerController = VideoPlayerController.network(sourceUrl);
+
+    controller = ChewieController(
+      aspectRatio: 3 / 2,
+      autoPlay: false,
+      looping: true,
+      autoInitialize: true,
+      placeholder: postUrl != null ? Center(child: Image.network(postUrl)) : SizedBox(),
+      videoPlayerController: videoPlayerController,
+    );
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    controller.pause();
+    controller.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return SeparatorElement(
+        adaptiveMap: adaptiveMap,
+        child: Chewie(
+          controller: controller,
+        ));
+  }
+}
