+++ lib/src/flutter_adaptive_cards.dart
@@ -1,18 +1,17 @@
 library flutter_adaptive_cards;
 
 import 'dart:async';
-import 'package:flutter_adaptive_cards/src/action_handler.dart';
-import 'package:flutter_adaptive_cards/src/registry.dart';
-import 'package:uuid/uuid.dart';
 import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_adaptive_cards/src/elements/input.dart';
+import 'package:flutter_adaptive_cards/src/action_handler.dart';
+import 'package:flutter_adaptive_cards/src/registry.dart';
 import 'package:flutter_adaptive_cards/src/utils.dart';
 import 'package:http/http.dart' as http;
 import 'package:provider/provider.dart';
 
-import 'elements/base.dart';
+import 'base.dart';
 
 abstract class AdaptiveCardContentProvider {
   AdaptiveCardContentProvider({@required this.hostConfigPath});
@@ -28,8 +27,7 @@ abstract class AdaptiveCardContentProvider {
 }
 
 class MemoryAdaptiveCardContentProvider extends AdaptiveCardContentProvider {
-  MemoryAdaptiveCardContentProvider(
-      {@required this.content, @required String hostConfigPath})
+  MemoryAdaptiveCardContentProvider({@required this.content, @required String hostConfigPath})
       : super(hostConfigPath: hostConfigPath);
 
   Map content;
@@ -41,8 +39,7 @@ class MemoryAdaptiveCardContentProvider extends AdaptiveCardContentProvider {
 }
 
 class AssetAdaptiveCardContentProvider extends AdaptiveCardContentProvider {
-  AssetAdaptiveCardContentProvider(
-      {@required this.path, @required String hostConfigPath})
+  AssetAdaptiveCardContentProvider({@required this.path, @required String hostConfigPath})
       : super(hostConfigPath: hostConfigPath);
 
   String path;
@@ -54,8 +51,7 @@ class AssetAdaptiveCardContentProvider extends AdaptiveCardContentProvider {
 }
 
 class NetworkAdaptiveCardContentProvider extends AdaptiveCardContentProvider {
-  NetworkAdaptiveCardContentProvider(
-      {@required this.url, @required String hostConfigPath})
+  NetworkAdaptiveCardContentProvider({@required this.url, @required String hostConfigPath})
       : super(hostConfigPath: hostConfigPath);
 
   String url;
@@ -67,7 +63,6 @@ class NetworkAdaptiveCardContentProvider extends AdaptiveCardContentProvider {
 }
 
 class AdaptiveCard extends StatefulWidget {
-
   AdaptiveCard({
     Key key,
     @required this.adaptiveCardContentProvider,
@@ -89,8 +84,7 @@ class AdaptiveCard extends StatefulWidget {
     this.onOpenUrl,
     this.showDebugJson = true,
     this.approximateDarkThemeColors = true,
-  }) : adaptiveCardContentProvider = NetworkAdaptiveCardContentProvider(
-            url: url, hostConfigPath: hostConfigPath);
+  }) : adaptiveCardContentProvider = NetworkAdaptiveCardContentProvider(url: url, hostConfigPath: hostConfigPath);
 
   AdaptiveCard.asset({
     Key key,
@@ -102,8 +96,7 @@ class AdaptiveCard extends StatefulWidget {
     this.onOpenUrl,
     this.showDebugJson = true,
     this.approximateDarkThemeColors = true,
-  }) : adaptiveCardContentProvider = AssetAdaptiveCardContentProvider(
-            path: assetPath, hostConfigPath: hostConfigPath);
+  }) : adaptiveCardContentProvider = AssetAdaptiveCardContentProvider(path: assetPath, hostConfigPath: hostConfigPath);
 
   AdaptiveCard.memory({
     Key key,
@@ -115,8 +108,8 @@ class AdaptiveCard extends StatefulWidget {
     this.onOpenUrl,
     this.showDebugJson = true,
     this.approximateDarkThemeColors = true,
-  }) : adaptiveCardContentProvider = MemoryAdaptiveCardContentProvider(
-            content: content, hostConfigPath: hostConfigPath);
+  }) : adaptiveCardContentProvider =
+            MemoryAdaptiveCardContentProvider(content: content, hostConfigPath: hostConfigPath);
 
   final AdaptiveCardContentProvider adaptiveCardContentProvider;
 
@@ -134,7 +127,6 @@ class AdaptiveCard extends StatefulWidget {
 }
 
 class _AdaptiveCardState extends State<AdaptiveCard> {
-
   Map map;
   Map hostConfig;
 
@@ -143,22 +135,19 @@ class _AdaptiveCardState extends State<AdaptiveCard> {
   Function(Map map) onSubmit;
   Function(String url) onOpenUrl;
 
-
   @override
   void initState() {
     super.initState();
     widget.adaptiveCardContentProvider.loadHostConfig().then((hostConfigMap) {
       setState(() {
-        if(mounted) {
+        if (mounted) {
           hostConfig = hostConfigMap;
         }
       });
     });
-    widget.adaptiveCardContentProvider
-        .loadAdaptiveCardContent()
-        .then((adaptiveMap) {
+    widget.adaptiveCardContentProvider.loadAdaptiveCardContent().then((adaptiveMap) {
       setState(() {
-        if(mounted) {
+        if (mounted) {
           map = adaptiveMap;
         }
       });
@@ -168,46 +157,42 @@ class _AdaptiveCardState extends State<AdaptiveCard> {
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    if(widget.cardRegistry != null) {
+    if (widget.cardRegistry != null) {
       cardRegistry = widget.cardRegistry;
     } else {
       CardRegistry cardRegistry = DefaultCardRegistry.of(context);
-      if(cardRegistry != null) {
+      if (cardRegistry != null) {
         this.cardRegistry = cardRegistry;
       } else {
         this.cardRegistry = const CardRegistry();
       }
     }
 
-
-    if(widget.onSubmit != null) {
+    if (widget.onSubmit != null) {
       onSubmit = widget.onSubmit;
     } else {
       var foundOnSubmit = DefaultAdaptiveCardHandlers.of(context)?.onSubmit;
-      if(foundOnSubmit != null) {
+      if (foundOnSubmit != null) {
         onSubmit = foundOnSubmit;
       } else {
         onSubmit = (it) {
-          Scaffold.of(context).showSnackBar(SnackBar(
-              content: Text("No handler found for: \n" + it.toString())));
+          Scaffold.of(context).showSnackBar(SnackBar(content: Text("No handler found for: \n" + it.toString())));
         };
       }
     }
 
-    if(widget.onOpenUrl != null) {
+    if (widget.onOpenUrl != null) {
       onOpenUrl = widget.onOpenUrl;
     } else {
       var foundOpenUrl = DefaultAdaptiveCardHandlers.of(context)?.onOpenUrl;
-      if(foundOpenUrl  != null) {
+      if (foundOpenUrl != null) {
         onOpenUrl = foundOpenUrl;
       } else {
         onOpenUrl = (it) {
-          Scaffold.of(context).showSnackBar(SnackBar(
-              content: Text("No handler found for: \n" + it.toString())));
+          Scaffold.of(context).showSnackBar(SnackBar(content: Text("No handler found for: \n" + it.toString())));
         };
       }
     }
-
   }
 
   @override
@@ -215,7 +200,9 @@ class _AdaptiveCardState extends State<AdaptiveCard> {
     if (map == null || hostConfig == null) {
       return widget.placeholder ?? const SizedBox();
     }
-    return RawAdaptiveCard.fromMap(map, hostConfig,
+    return RawAdaptiveCard.fromMap(
+      map,
+      hostConfig,
       cardRegistry: cardRegistry,
       onOpenUrl: onOpenUrl,
       onSubmit: onSubmit,
@@ -225,15 +212,15 @@ class _AdaptiveCardState extends State<AdaptiveCard> {
   }
 }
 
-
-
 /// Main entry point to adaptive cards.
 ///
 /// This widget takes a [map] (which usually is just a json decoded string) and
 /// displays in natively. Additionally a host config needs to be provided for
 /// styling.
 class RawAdaptiveCard extends StatefulWidget {
-  RawAdaptiveCard.fromMap(this.map, this.hostConfig, {
+  RawAdaptiveCard.fromMap(
+    this.map,
+    this.hostConfig, {
     this.cardRegistry = const CardRegistry(),
     @required this.onSubmit,
     @required this.onOpenUrl,
@@ -264,10 +251,10 @@ class RawAdaptiveCardState extends State<RawAdaptiveCard> {
   // The root element
   Widget _adaptiveElement;
 
-
   static RawAdaptiveCardState of(BuildContext context) {
     return Provider.of<RawAdaptiveCardState>(context);
   }
+
   @override
   void initState() {
     super.initState();
@@ -277,8 +264,7 @@ class RawAdaptiveCardState extends State<RawAdaptiveCard> {
     idGenerator = UUIDGenerator();
     cardRegistry = widget.cardRegistry;
 
-    _adaptiveElement =
-        widget.cardRegistry.getElement(widget.map);
+    _adaptiveElement = widget.cardRegistry.getElement(widget.map);
   }
 
   /// Every widget can access method of this class, meaning setting the state
@@ -287,17 +273,13 @@ class RawAdaptiveCardState extends State<RawAdaptiveCard> {
     setState(() {});
   }
 
-
-
   /// Submits all the inputs of this adaptive card, does it by recursively
   /// visiting the elements in the tree
   void submit(Map map) {
-
-
     var visitor;
     visitor = (element) {
-      if(element is StatefulElement) {
-        if(element.state is AdaptiveInputMixin) {
+      if (element is StatefulElement) {
+        if (element.state is AdaptiveInputMixin) {
           (element.state as AdaptiveInputMixin).appendInput(map);
         }
       }
@@ -336,7 +318,7 @@ class RawAdaptiveCardState extends State<RawAdaptiveCard> {
     Widget child = _adaptiveElement;
 
     assert(() {
-      if(widget.showDebugJson) {
+      if (widget.showDebugJson) {
         child = Column(
           children: <Widget>[
             FlatButton(
@@ -344,26 +326,30 @@ class RawAdaptiveCardState extends State<RawAdaptiveCard> {
               onPressed: () {
                 JsonEncoder encoder = new JsonEncoder.withIndent('  ');
                 String prettyprint = encoder.convert(widget.map);
-                showDialog(context: context, builder: (context) {
-                  return AlertDialog(
-                    title: Text("JSON (only added in debug mode, you can also turn"
-                        "it of manually by passing showDebugJson = false)"),
-                    content: SingleChildScrollView(child: Text(prettyprint)),
-                    actions: <Widget>[
-                      Center(
-                        child: FlatButton(
-                          onPressed: () => Navigator.of(context).pop(),
-                          child: Text("Thanks"),
-                        ),
-                      )
-                    ],
-                    contentPadding: EdgeInsets.all(8.0),
-                  );
-                });
+                showDialog(
+                    context: context,
+                    builder: (context) {
+                      return AlertDialog(
+                        title: Text("JSON (only added in debug mode, you can also turn"
+                            "it of manually by passing showDebugJson = false)"),
+                        content: SingleChildScrollView(child: Text(prettyprint)),
+                        actions: <Widget>[
+                          Center(
+                            child: FlatButton(
+                              onPressed: () => Navigator.of(context).pop(),
+                              child: Text("Thanks"),
+                            ),
+                          )
+                        ],
+                        contentPadding: EdgeInsets.all(8.0),
+                      );
+                    });
               },
               child: Text("Debug show the JSON"),
             ),
-            Divider(height: 0,),
+            Divider(
+              height: 0,
+            ),
             child,
           ],
         );
@@ -416,9 +402,7 @@ typedef AdaptiveElementVisitor = void Function(AdaptiveElement element);
 /// implement the method [visitChildren] and call visitor(this) in addition call
 /// [visitChildren] on each child with the passed visitor.
 abstract class AdaptiveElement {
-  AdaptiveElement(
-      {@required this.adaptiveMap,
-      @required this.widgetState}) {
+  AdaptiveElement({@required this.adaptiveMap, @required this.widgetState}) {
     loadTree();
   }
 
@@ -479,16 +463,11 @@ abstract class AdaptiveElement {
 
   @override
   bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is AdaptiveElement &&
-          runtimeType == other.runtimeType &&
-          id == other.id;
+      identical(this, other) || other is AdaptiveElement && runtimeType == other.runtimeType && id == other.id;
 
   @override
   int get hashCode => id.hashCode;
 }
-
-
 
 /// Resolves values based on the host config.
 ///
@@ -497,20 +476,20 @@ class ReferenceResolver {
   ReferenceResolver({
     this.hostConfig,
     this.currentStyle,
-});
+  });
 
   final Map hostConfig;
 
   final String currentStyle;
 
   dynamic resolve(String key, String value) {
-    dynamic res =  hostConfig[key][firstCharacterToLowerCase(value)];
+    dynamic res = hostConfig[key][firstCharacterToLowerCase(value)];
     assert(res != null, "Could not find hostConfig[$key][${firstCharacterToLowerCase(value)}]");
     return res;
   }
 
   dynamic get(String key) {
-    dynamic res =  hostConfig[key];
+    dynamic res = hostConfig[key];
     assert(res != null, "Could not find hostConfig[$key]");
     return res;
   }
@@ -523,8 +502,8 @@ class ReferenceResolver {
         "FontWeight '${value ?? "default"}' was not found in the host_config. \n\n"
         "The available font weights were: \n\n"
         "${(hostConfig["fontWeights"] as Map).entries.map((entry) => "${entry.key}: ${entry.value}\n").toList()}");
-    FontWeight fontWeight = FontWeight.values.firstWhere(
-        (possibleWeight) => possibleWeight.toString() == "FontWeight.w$weight");
+    FontWeight fontWeight =
+        FontWeight.values.firstWhere((possibleWeight) => possibleWeight.toString() == "FontWeight.w$weight");
     assert(fontWeight != null, "There is no FontWight.w$weight");
     return fontWeight;
   }
@@ -555,12 +534,10 @@ class ReferenceResolver {
     String subtleOrDefault = isSubtle ?? false ? "subtle" : "default";
     final style = currentStyle ?? "default";
     // Make it case insensitive
-    String colorValue = hostConfig["containerStyles"][style]
-            ["foregroundColors"][firstCharacterToLowerCase(myColor)]
-        [subtleOrDefault];
+    String colorValue =
+        hostConfig["containerStyles"][style]["foregroundColors"][firstCharacterToLowerCase(myColor)][subtleOrDefault];
     return parseColor(colorValue);
   }
-
 
   ReferenceResolver copyWith({String style}) {
     assert(style == null || style == "default" || style == "emphasis");
@@ -579,8 +556,3 @@ class ReferenceResolver {
     return intSpacing.toDouble();
   }
 }
-
-
-
-
-
