+++ lib/src/registry.dart
@@ -1,13 +1,28 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_adaptive_cards/src/elements/basics.dart';
-import 'package:flutter_adaptive_cards/src/elements/actions.dart';
-import 'package:flutter_adaptive_cards/src/elements/input.dart';
-import 'package:flutter_adaptive_cards/src/flutter_adaptive_cards.dart';
 
-import 'elements/base.dart';
+import '../flutter_adaptive_cards.dart';
+import 'adaptive_card_element.dart';
+import 'base.dart';
+import 'containers/column_set.dart';
+import 'containers/container.dart';
+import 'containers/fact_set.dart';
+import 'containers/image_set.dart';
+import 'elements/action_set.dart';
+import 'elements/actions/open_url.dart';
+import 'elements/actions/show_card.dart';
+import 'elements/actions/sumbit.dart';
+import 'elements/image.dart';
+import 'elements/media.dart';
+import 'elements/text_block.dart';
+import 'elements/unknown.dart';
+import 'inputs/choise_set.dart';
+import 'inputs/date.dart';
+import 'inputs/number.dart';
+import 'inputs/text.dart';
+import 'inputs/time.dart';
+import 'inputs/toggle.dart';
 
-typedef ElementCreator = Widget Function(
-    Map<String, dynamic> map);
+typedef ElementCreator = Widget Function(Map<String, dynamic> map);
 
 /// Entry point for registering adaptive cards
 ///
@@ -24,10 +39,7 @@ typedef ElementCreator = Widget Function(
 /// Delete an element even if you have provided it yourself via the [addedElements]
 ///
 class CardRegistry {
-  const CardRegistry(
-      {this.removedElements = const [],
-      this.addedElements = const {},
-      this.addedActions = const {}});
+  const CardRegistry({this.removedElements = const [], this.addedElements = const {}, this.addedActions = const {}});
 
   /// Provide custom elements to use.
   /// When providing an element which is already defined, it is overwritten
@@ -38,15 +50,14 @@ class CardRegistry {
   /// Remove specific elements from the list
   final List<String> removedElements;
 
-  Widget getElement(
-      Map<String, dynamic> map) {
+  Widget getElement(Map<String, dynamic> map) {
     String stringType = map["type"];
 
     if (removedElements.contains(stringType))
-    return AdaptiveUnknown(
-      type: stringType,
-      adaptiveMap: map,
-    );
+      return AdaptiveUnknown(
+        type: stringType,
+        adaptiveMap: map,
+      );
 
     if (addedElements.containsKey(stringType)) {
       return addedElements[stringType](map);
@@ -63,7 +74,7 @@ class CardRegistry {
         assert(false, "Action.ShowCard can only be used directly by the root card");
         return null;
       case "Action.OpenUrl":
-        return GenericActionOpenUrl(map,state);
+        return GenericActionOpenUrl(map, state);
       case "Action.Submit":
         return GenericSubmitAction(map, state);
     }
@@ -71,9 +82,7 @@ class CardRegistry {
     return null;
   }
 
-
-  Widget getAction(
-      Map<String, dynamic> map) {
+  Widget getAction(Map<String, dynamic> map) {
     String stringType = map["type"];
 
     if (removedElements.contains(stringType))
@@ -92,27 +101,38 @@ class CardRegistry {
   /// This returns an [AdaptiveElement] with the correct type.
   ///
   /// It looks at the [type] property and decides which object to construct
-  Widget _getBaseElement(
-      Map<String, dynamic> map) {
+  Widget _getBaseElement(Map<String, dynamic> map) {
     String stringType = map["type"];
 
     switch (stringType) {
       case "Media":
         return AdaptiveMedia(adaptiveMap: map);
       case "Container":
-        return AdaptiveContainer(adaptiveMap: map,);
+        return AdaptiveContainer(
+          adaptiveMap: map,
+        );
       case "TextBlock":
-        return AdaptiveTextBlock(adaptiveMap: map,);
+        return AdaptiveTextBlock(
+          adaptiveMap: map,
+        );
       case "ActionSet":
         return ActionSet(adaptiveMap: map);
       case "AdaptiveCard":
-        return AdaptiveCardElement(adaptiveMap: map,);
+        return AdaptiveCardElement(
+          adaptiveMap: map,
+        );
       case "ColumnSet":
-        return AdaptiveColumnSet(adaptiveMap: map,);
+        return AdaptiveColumnSet(
+          adaptiveMap: map,
+        );
       case "Image":
-        return AdaptiveImage(adaptiveMap: map,);
+        return AdaptiveImage(
+          adaptiveMap: map,
+        );
       case "FactSet":
-        return AdaptiveFactSet(adaptiveMap: map,);
+        return AdaptiveFactSet(
+          adaptiveMap: map,
+        );
       case "ImageSet":
         return AdaptiveImageSet(adaptiveMap: map);
       case "Input.Text":
@@ -120,30 +140,51 @@ class CardRegistry {
       case "Input.Number":
         return AdaptiveNumberInput(adaptiveMap: map);
       case "Input.Date":
-        return AdaptiveDateInput(adaptiveMap: map,);
+        return AdaptiveDateInput(
+          adaptiveMap: map,
+        );
       case "Input.Time":
-        return AdaptiveTimeInput(adaptiveMap: map,);
+        return AdaptiveTimeInput(
+          adaptiveMap: map,
+        );
       case "Input.Toggle":
-        return AdaptiveToggle(adaptiveMap: map,);
+        return AdaptiveToggle(
+          adaptiveMap: map,
+        );
       case "Input.ChoiceSet":
-        return AdaptiveChoiceSet(adaptiveMap: map,);
+        return AdaptiveChoiceSet(
+          adaptiveMap: map,
+        );
     }
-    return AdaptiveUnknown(adaptiveMap: map, type: stringType,);
+    return AdaptiveUnknown(
+      adaptiveMap: map,
+      type: stringType,
+    );
   }
 
   Widget _getBaseAction(
-      Map<String, dynamic> map,) {
+    Map<String, dynamic> map,
+  ) {
     String stringType = map["type"];
 
     switch (stringType) {
       case "Action.ShowCard":
-        return AdaptiveActionShowCard(adaptiveMap: map,);
+        return AdaptiveActionShowCard(
+          adaptiveMap: map,
+        );
       case "Action.OpenUrl":
-        return AdaptiveActionOpenUrl(adaptiveMap: map,);
+        return AdaptiveActionOpenUrl(
+          adaptiveMap: map,
+        );
       case "Action.Submit":
-        return AdaptiveActionSubmit(adaptiveMap: map,);
+        return AdaptiveActionSubmit(
+          adaptiveMap: map,
+        );
     }
-    return AdaptiveUnknown(adaptiveMap: map, type: stringType,);
+    return AdaptiveUnknown(
+      adaptiveMap: map,
+      type: stringType,
+    );
   }
 }
 
@@ -158,9 +199,10 @@ class DefaultCardRegistry extends InheritedWidget {
 
   static CardRegistry of(BuildContext context) {
     DefaultCardRegistry cardRegistry = context.inheritFromWidgetOfExactType(DefaultCardRegistry);
-    if(cardRegistry == null) return null;
+    if (cardRegistry == null) return null;
     return cardRegistry.cardRegistry;
   }
+
   @override
   bool updateShouldNotify(DefaultCardRegistry oldWidget) => oldWidget != this;
 }
